Clarifying locations README.md
I thought it would be helpful to users of the addon that don't yet know about the Obsidian file structure to have a note on where to add the configs (especially since the locations are somewhat hidden in the official docs).  I hope I've added the note in an  appropriate place.